learnsign/server/requirements.txt
flask==2.3.0
flask-cors==4.0.0
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.0.0
mediapipe>=0.10.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
gunicorn>=21.0.0
mediapipe>=0.10.0
opencv-python>=4.8.0
